+++ NEWS.md
@@ -1,4 +1,41 @@
+# MoNAn 1.0.0
+
+!! Backward breaking change !!
+
+The "cache" is not visible to the user anymore, it is created automatically
+and there is no more use for the function createWeightedCache.
+
+The following functions are used differently:
+
+- gofMobilityNetwork does not require the specification of simulations anymore!
+
+New functions (aliases) for easier use
+
+- monanDependent (createEdgelist), monanNodes (createNodeSet), 
+  monanEdges (createNodeSet), monadicCovariate (createNodeAttribute),
+  dyadicCovariate (createNetwork), monanEstimate (estimateMobilityNetwork),
+  monanSimulate (simulateMobilityNetwork), monanGOF (gofMobilityNetwork)
+
+Further changes:
+
+- createEffect and addEffects; specification of node.attribute and
+  edge.attribute instead of attribute.index and resource.attribute.index
+
+- The results objects are much smaller and contain less redundant information.
+
+- The State and Effects objects have print functions for easier inspection.
+
+- The algorithm is now stored in the outcome object for later reference.
+
 # MoNAn 0.3.0
+
+New functions to specify the model: createEffects and addEffect
+New function to generate the process state: monanDataCreate
+
+Creation of alias: monanAlgorithmCreate for createAlgorithm
+Creation of alias: monan07 for estimateMobilityNetwork
+
+Implementation of new effects
 
 # MoNAn 0.2.0